Move `RedirectHeadlessPreviewMixin` to home app
This is a more central location for a mixin we want to apply to models
in several different apps.
