+++ report/models.py
@@ -18,41 +18,14 @@ from wagtail.documents.blocks import DocumentChooserBlock
 from wagtail.documents.edit_handlers import DocumentChooserPanel
 from wagtail.images.edit_handlers import ImageChooserPanel
 from wagtail.search import index
-from wagtail_headless_preview.models import HeadlessPreviewMixin, PagePreview
 
-from home.models import AbstractHomeContentPage, Post
+from home.models import AbstractHomeContentPage, Post, RedirectHeadlessPreviewMixin
 from programs.models import AbstractContentPage
 
 from .blocks import EndnoteBlock, FeaturedReportSectionBlock, ReportSectionBlock
 from .tasks import (
     generate_pdf, generate_report_contents, get_report_authors, parse_pdf, write_pdf,
 )
-
-
-class RedirectHeadlessPreviewMixin(HeadlessPreviewMixin):
-    def serve_preview(self, request, mode_name):
-        use_live_preview = request.GET.get("live_preview")
-        token = request.COOKIES.get("used-token")
-
-        if use_live_preview and token:
-            page_preview, existed = self.update_page_preview(token)
-            PagePreview.garbage_collect()
-
-            from wagtail_headless_preview.signals import (
-                preview_update,
-            )  # Imported locally as live preview is optional
-
-            preview_update.send(sender=HeadlessPreviewMixin, token=token)
-        else:
-            PagePreview.garbage_collect()
-            page_preview = self.create_page_preview()
-            page_preview.save()
-
-        response_token = token or page_preview.token
-
-        response = redirect(self.get_preview_url(response_token))  # from django.shortcuts import redirect
-
-        return response
 
 
 class Report(RedirectHeadlessPreviewMixin, RoutablePageMixin, Post):
